move fido2 related files into fido2-dir
